requirements.txt
instaloader
Gooey
yt-dlp
PyYAML
pytest
colored==1.4.4
